+++ R/clusterSim.R
@@ -5,7 +5,8 @@
 ##' @title clusterSim
 ##' @param cluster1 a vector of gene IDs
 ##' @param cluster2 another vector of gene IDs
-##' @param organism one of "hsa" and "mmu"
+##' @param organism species of the Entrez gene IDs. If omitted, it is inferred
+##' from `ont`.
 ##' @param ont one of "HDO", "HPO" and "MPO"
 ##' @param measure One of "Resnik", "Lin", "Rel", "Jiang" and "Wang" methods.
 ##' @param combine One of "max", "avg", "rcmax", "BMA" methods, for combining
@@ -22,13 +23,17 @@
 clusterSim <- function(cluster1, 
                        cluster2, 
                        ont = "HDO",
-                       organism = "hsa",
+                       organism = NULL,
                        measure="Wang", 
                        combine="BMA") {
-    if (ont == "DO") ont <- 'HDO'
+    info <- .resolve_ontology_organism(ont, organism)
+    ont <- info$ontology
+    organism <- info$organism
+    .validate_entrez_ids(cluster1, "cluster1")
+    .validate_entrez_ids(cluster2, "cluster2")
 
-    do1 <- sapply(cluster1, gene2DO, organism = organism)
-    do2 <- sapply(cluster2, gene2DO, organism = organism)
+    do1 <- sapply(cluster1, gene2DO, organism = organism, ont = ont)
+    do2 <- sapply(cluster2, gene2DO, organism = organism, ont = ont)
 
     do1 <- unlist(do1)
     do2 <- unlist(do2)