+++ R/enrichNCG.R
@@ -9,7 +9,6 @@
 #' @return A \code{enrichResult} instance
 #' @export
 #' @author Guangchuang Yu
-#' @importFrom utils read.delim
 enrichNCG <- function(gene,
                       pvalueCutoff = 0.05,
                       pAdjustMethod = "BH",
@@ -31,8 +30,7 @@ enrichNCG <- function(gene,
 }
 
 get_NCG_data <- function() {
-    if (!exists(".DOSEenv")) .initial()
-    .DOSEEnv <- get(".DOSEEnv", envir = .GlobalEnv)
+    .DOSEEnv <- get_dose_env()
     
     if (exists(".NCG_DOSE_GSON", envir=.DOSEEnv)) {
         res <- get(".NCG_DOSE_GSON", envir = .DOSEEnv)
@@ -42,29 +40,9 @@ get_NCG_data <- function() {
     urls <- c("https://yulab-smu.top/DOSE",
               "https://raw.githubusercontent.com/YuLab-SMU/DOSE/refs/heads/gh-pages")
     
-    dbfile <- yulab.utils::download_yulab_file("NCG.tsv.gz", urls, 
-                                               gzfile = FALSE, appname = "DOSE")
-    
-    ncg <- read.delim(gzfile(dbfile), stringsAsFactors = FALSE)
-    PATHID2EXTID <- split(as.character(ncg$entrez), as.character(ncg$cancer_type))
-
-    # gsid2gene
-    gsid2gene <- stack(PATHID2EXTID)
-    colnames(gsid2gene) <- c("gene", "gsid")
-    gsid2gene <- gsid2gene[, c("gsid", "gene")]
-
-    # gsid2name: use cancer_type as both ID and name
-    gsid2name <- data.frame(gsid = unique(ncg$cancer_type),
-                            name = unique(ncg$cancer_type),
-                            stringsAsFactors = FALSE)
-
-    gson_obj <- gson::gson(gsid2gene = gsid2gene, 
-                           gsid2name = gsid2name,
-                           species = "Homo sapiens",
-                           gsname = "NCG",
-                           keytype = "ENTREZID",
-                           version = "unknown",
-                           accessed_date = as.character(Sys.Date()))
+    dbfile <- yulab.utils::download_yulab_file("NCG.gson", urls, 
+                                               gzfile = TRUE, appname = "DOSE")
+    gson_obj <- gson::read.gson(dbfile)
 
     assign(".NCG_DOSE_GSON", gson_obj, envir = .DOSEEnv)
     return(gson_obj)