+++ data-raw/generate-md5.R
@@ -1,5 +1,5 @@
 ff <- list.files(pattern = ".sqlite$")
-ff2 <- list.files(pattern = "\\.tsv\\.gz$")
+ff2 <- list.files(pattern = "\\.(tsv|gson)\\.gz$")
 md5 <- vapply(ff, function(f) digest::digest(f, algo='md5', file=TRUE), character(1))
 md5_tsv <- vapply(ff2, function(f) digest::digest(f, algo='md5', file=TRUE), character(1))
 md5 <- c(md5, md5_tsv)